+++ tests/tests/src/files_suite-commands.adb
@@ -624,8 +624,9 @@ package body Files_Suite.Commands is
       Shift (Files.Types.Shift_Key) := True;
       Ctrl_Shift (Files.Types.Control_Key) := True;
       Ctrl_Shift (Files.Types.Shift_Key) := True;
-      Assert (Files.Commands.Command_Count = 70, "all expected commands are registered");
+      Assert (Files.Commands.Command_Count = 71, "all expected commands are registered");
       Assert (Files.Commands.Contains ("navigate.recent"), "recent-view command identifier is registered");
+      Assert (Files.Commands.Contains ("search.contents"), "content-search command identifier is registered");
       Assert (Files.Commands.Contains ("recent.clear"), "clear-recent command identifier is registered");
       Assert (Files.Commands.Contains ("label.set"), "set-color-label command identifier is registered");
       Assert (Files.Commands.Contains ("edit.copy_path"), "copy-path command identifier is registered");